apps/api/pyproject.toml
[project]
name = "creator-cfo-api"
version = "0.1.0"
description = "FastAPI service for the Creator CFO monorepo"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115,<1.0",
    "pydantic-settings>=2.6,<3.0",
    "pyyaml>=6.0,<7.0",
    "uvicorn[standard]>=0.32,<1.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28,<1.0",
    "pytest>=8.3,<9.0",
    "ruff>=0.8,<1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/creator_cfo_api"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
src = ["src", "tests", "scripts"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

